:truck: move FindTargetProfileSummariesForTraining usecase
